requirements.txt
fastapi[all]
uvicorn
pillow
PyMuPDF
python-multipart